pyproject.toml
[project]
name = "desi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pyedifice>=4.4.1",
    "pyqt6>=6.10.0",
    "pyside6-essentials>=6.10.0",
    "watchdog>=6.0.0",
]
